refactor stores/tags.ts, extract api and util
